RavenDB-17793 : address PR comments and fix errors after rebase
